requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\goros\Desktop\Discretas_bot\bot.py: 4
beautifulsoup4 == 4.10.0

# C:\Users\goros\Desktop\Discretas_bot\bot.py: 10
matplotlib == 3.5.0

# C:\Users\goros\Desktop\Discretas_bot\bot.py: 8
networkx == 2.6.3

# C:\Users\goros\Desktop\Discretas_bot\bot.py: 3,11
python_telegram_bot == 13.8.1

# C:\Users\goros\Desktop\Discretas_bot\bot.py: 5
requests == 2.26.0
